+++ scripts/release.js
@@ -1,0 +1,11 @@
+// 研发流程
+// feature分支从Master拉
+// dev分支为预发布分支
+// dev分支验证完成后在master分支进行发布，修改Changelog
+
+// beta release
+// beta版在feature上发,
+// changelog体现新版本，有unreleased
+// build => .... => publish
+
+// main release