+++ src/client/components/SandBox/styled.ts
@@ -1,0 +1,8 @@
+
+import styled from 'styled-components'
+
+export const StyledSandBox = styled.div`
+  :hover .bottom-tools {
+    visibility: visible;
+  }
+`