+++ src/client/styled/vars.ts
@@ -2,9 +2,9 @@ import { createGlobalStyle } from "styled-components";
 
 export const CssVariables = createGlobalStyle`
   :root {
-    --sidebar-width: 20em;
+    --sidebar-width: 15em;
     --header-height: 4rem;
-    --toc-width: 280px;
+    --toc-width: 15em;
 
     --font-size: 16px;
     --font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
@@ -31,7 +31,7 @@ export const CssVariables = createGlobalStyle`
     --c-code-bg: #ececec;
     --c-code-keyword: #89b482;
     --c-code-function: #a9b665;
-    --c-code-attr: #fff;
+    --c-code-attr: #D8A657;
     --c-code-string: #D8A657;
     --c-code-built-in: #7DAEA3;
     --c-code-name: #e78a4e;